Split MlnFfiMapSession into a map core and a render session
MlnFfiMapCore owns the loop thread's half: the runtime loop, requested
style and camera, the mirrored viewport, and the style binding.
MlnFfiMapSession keeps the renderer-thread half: the host session,
render-session handle lifecycle, and frame pacing, driving the core
through a narrow render-access seam that answers queries empty when no
session is attached. Both still live and die with the composition.
