+++ lib/maplibre-compose/src/maplibreNativeTest/kotlin/org/maplibre/compose/sources/GeoJsonSourceAttachWaitTest.kt
@@ -39,7 +39,7 @@ class GeoJsonSourceAttachWaitTest {
       val entered = TestLatch(1)
       val release = TestLatch(1)
       assertTrue(
-        fixture.session.postOwnerTaskForTest {
+        fixture.core.postOwnerTaskForTest {
           entered.countDown()
           check(release.await(5_000))
         }
